refactors post edits
